+++ src/input/timeScroller.ts
@@ -1,12 +1,15 @@
 import m, { CVnode, ClassComponent } from "mithril";
+import { IConfig } from "../interface/config";
 import { TPropStream } from "../interface/widget";
 import { theme } from "../theme";
+import { getConfig, getIcon } from "../config";
 
 interface ITimeScroller {
 	readonly value: TPropStream;
 	readonly min: number;
 	readonly max: number;
 	readonly step?: number;
+	readonly config?: Partial<IConfig>;
 }
 
 export class TimeScroller implements ClassComponent<ITimeScroller> {
@@ -27,17 +30,21 @@ export class TimeScroller implements ClassComponent<ITimeScroller> {
 		value(clamped.toString());
 	}
 
-	public view({ attrs: { value, step = 1, min, max } }: CVnode<ITimeScroller>) {
+	public view({ attrs: { value, step = 1, min, max, config } }: CVnode<ITimeScroller>) {
 		return m(".flex.flex-column.items-center", [
-			m(".fas.fa-chevron-up.pointer", {
+			m(".pointer", {
 				onclick: () => TimeScroller.applyStep(value, step, min, max)
-			}),
+			},
+				getIcon(getConfig("timeScrollerUpIcn", config), "")
+			),
 			m("span.mv2.w-100.mw-dd.tc.f6.fw6", {
 				class: theme.timeInputScrollerNumber
 			}, TimeScroller.formatValue(value)),
-			m(".fas.fa-chevron-down.pointer", {
+			m(".pointer", {
 				onclick: () => TimeScroller.applyStep(value, -step, min, max)
-			})
+			},
+				getIcon(getConfig("timeScrollerDownIcn", config), "")
+			)
 		]);
 	}